Notepad New Line Implementation
The Notepad now consists of Multiple Lines connected together using LinkedList.
The Pointer can now Move Up and Down the Lines.
The Display, Insertion, Deletion needed to be Updated

When a Character is Inserted or Deleted, the Links do not keep correct Sync with Up and Down Lines.
However, the Next and Previous Links Remain in Sync in each Line.